added all language files. bug fix for auth process
also improved performance of Apply by not fully refreshing the list with a request, but instead updating the element itself. The resorting my be looked more into for better performance results.

Currently there is stil bad performance when the list is to large
Also refreshing the list on page opening now for each page